+++ app/src/main/java/com/vireya/hydrocore/agenda/Agenda.java
@@ -1,4 +1,4 @@
-package com.vireya.hydrocore;
+package com.vireya.hydrocore.agenda;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.vireya.hydrocore.R;
 
 /**
  * A simple {@link Fragment} subclass.
