+++ app/src/main/java/com/vireya/hydrocore/entrada/EsqueceuSenha.java
@@ -1,4 +1,4 @@
-package com.vireya.hydrocore;
+package com.vireya.hydrocore.entrada;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.vireya.hydrocore.R;
 
 public class EsqueceuSenha extends AppCompatActivity {
