+++ app/src/main/java/com/vireya/hydrocore/entrada/Login.java
@@ -1,0 +1,34 @@
+package com.vireya.hydrocore.entrada;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.vireya.hydrocore.R;
+
+public class Login extends AppCompatActivity {
+
+    FirebaseAuth objAuth = FirebaseAuth.getInstance();
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        Button bt= findViewById(R.id.entrar);
+        bt.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+            }
+        });
+    }
+
+    }
+
+
+
+
+
