+++ app/src/main/res/layout/activity_esqueceu_senha.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EsqueceuSenha"
+    tools:context=".entrada.EsqueceuSenha"
     >
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
