+++ app/src/main/res/layout/activity_redefinir_senha.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RedefinirSenha">
+    tools:context=".entrada.RedefinirSenha">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
