+++ app/src/main/res/layout/fragment_tarefas.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Tarefas">
+    tools:context=".tarefas.Tarefas">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
